+++ engine/src/layer.h
@@ -23,8 +23,8 @@ typedef struct {
 
 	shrlayer **layers;
 
-	size_t used;
-	size_t size;
+	usize used;
+	usize size;
 
 } shrlayer_stack;
 
@@ -34,7 +34,7 @@ void shrlayer_stack_destory(shrlayer_stack *stack);
 void shrlayer_stack_push(shrlayer_stack *stack, shrlayer *layer);
 void shrlayer_stack_push_front(shrlayer_stack *stack, shrlayer *layer);
 void shrlayer_stack_pop(shrlayer_stack *stack, shrlayer *layer);
-shrlayer* shrlayer_stack_get(shrlayer_stack *stack, int index);
+shrlayer* shrlayer_stack_get(shrlayer_stack *stack, usize index);
 int shrlayer_stack_size(shrlayer_stack *stack);
 
 #endif // !LAYER_H
